requirements.txt
streamlit
reportlab
pillow
google-generativeai